requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.30
SQLAlchemy>=2.0
psycopg2-binary>=2.9
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
bcrypt>=4.1
argon2-cffi>=23.1
jinja2>=3.1
pydantic>=2.6
email-validator>=2.1
httpx>=0.27
